refacto: backbone module
