+++ doctr/models/__init__.py
@@ -1,6 +1,5 @@
 from . import artefacts
-from .backbones import *
+from .classification import *
 from .detection import *
 from .recognition import *
 from .zoo import *
-from .classification import *
