--- doctr/models/backbones/__init__.py
@@ -1,4 +1,0 @@
-from .mobilenet import *
-from .resnet import *
-from .vgg import *
-from .magc_resnet import *
