--- doctr/models/backbones/vgg/__init__.py
@@ -1,6 +1,0 @@
-from doctr.file_utils import is_tf_available, is_torch_available
-
-if is_tf_available():
-    from .tensorflow import *
-elif is_torch_available():
-    from .pytorch import *  # type: ignore[misc]
