+++ doctr/models/classification/__init__.py
@@ -1,2 +1,5 @@
 from .mobilenet import *
+from .resnet import *
+from .vgg import *
+from .magc_resnet import *
 from .zoo import *
