+++ doctr/models/classification/magc_resnet/__init__.py
