+++ doctr/models/classification/magc_resnet/pytorch.py
