+++ doctr/models/classification/magc_resnet/tensorflow.py
