+++ doctr/models/classification/resnet/pytorch.py
