+++ doctr/models/classification/resnet/tensorflow.py
