+++ doctr/models/classification/vgg/pytorch.py
