+++ doctr/models/classification/vgg/tensorflow.py
